gitLab integration fixes
Added gitlab4j-api integration in order to process gitlab push events.

BCIntegration got transformed into interface. Original code moved to GithubIntegrationImpl.

GitLabIntegrationImpl uses gitlab4j-api  model types.
