+++ pom.xml
@@ -43,7 +43,7 @@
       <dependency>
         <groupId>org.jboss.resteasy</groupId>
         <artifactId>resteasy-jaxrs</artifactId>
-        <version>3.0.19.Final</version>
+        <version>4.0.0.Beta5</version>
       </dependency>
       <dependency>
         <groupId>com.aliction</groupId>
@@ -75,6 +75,12 @@
         <groupId>me.porcelli.bc-git-integration</groupId>
         <artifactId>bc-git-integration-common</artifactId>
         <version>2.2.0-SNAPSHOT</version>
+      </dependency>
+          <!-- https://mvnrepository.com/artifact/org.gitlab4j/gitlab4j-api -->
+      <dependency>
+        <groupId>org.gitlab4j</groupId>
+        <artifactId>gitlab4j-api</artifactId>
+        <version>5.0.1</version>
       </dependency>
     </dependencies>
   </dependencyManagement>
